\name{BDPS}
\docType{data}
\alias{BDPS}
\title{Ring-width measurements of 10 Scots pine specimen sampled in Beutelsdorf, Germany}
\description{
  This data set provides measurements of secondary growth of 10 Scots pine specimen sampled
  in Beutelsdorf, Germany.
}
\usage{data(BDPS)}
\format{An rwl.data.frame containing ring-width measurements
 of ten Scots pine individuals from 1914-2016}
\keyword{datasets}