\name{chron.by.stem}
\alias{chron.by.stem}
\title{
Compute chronologies by stem
}
\description{
Computes individual based chronologies over all series belonging to a particular individual (as indicated by its ID).
}
\usage{
chron.by.stem(rwl, str.id1, str.id2)
}

\arguments{
  \item{rwl}{
rwl.data.frame such as generated by read.rwl from the dplR-package
}
  \item{str.id1}{
the first string of the columnames indicating the unique tree ID
}
  \item{str.id2}{
the last string of the columnames indicating the unique tree ID
}
}
\details{
make sure your column names allow for systematic identification of unique tree-IDs, e.g. for PAFP01E and PAFP01S str.id1=1,str.id2=7 would work
}
\value{
an rwl.data.frame containing ring-width series per individual as defined by their ID
}
\author{
Allan Buras
}


