\name{Clim.PA}
\docType{data}
\alias{Clim.PA}
\title{Climate data for the region southwest of Munich, Germany}
\description{
  This data set provides climate data representative for the region southwest of Munich.
}
\usage{data(Clim.PA)}
\format{A data.frame containing climate data for the region southwest of munich.}
\keyword{datasets}