\name{DBH.PA}
\docType{data}
\alias{DBH.PA}
\title{DBH measurements of 40 Picea abies specimen sampled southwest of Munich, Germany}
\description{
  This data set provides DBH measurements of 40 Picea abies specimen sampled southwest of Munich.
}
\usage{data(DBH.PA)}
\format{A data.frame containing DBH measurements used for the computation of basal area increments.}
\keyword{datasets}