\name{dendRolAB-package}
\alias{dendRolAB-package}
\alias{dendRolAB}
\docType{package}
\title{
\packageTitle{dendRolAB}
}
\description{
\packageDescription{dendRolAB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{
dendRolAB contains novel analytical tools for tree-ring scientists, such as the Principal Component Gradient Analysis (PCGA), the Bootstrapped Transfer Function Stability test (BTFS), and the Standardized Growth Change method (SGC).
}
}
\author{
\packageAuthor{dendRolAB}

Maintainer: \packageMaintainer{dendRolAB}
}
\references{
For further details see the corresponding video tutorials at dendroschool.org
}

