\name{SMCL}
\docType{data}
\alias{SMCL}
\title{Climate data representative for south of Munich.}
\description{
  This data set holds climate data representative for the region south of Munich, Germany, and was published in Rehschuh et al., 2017.
}
\usage{data(SMCL)}
\format{A data.frame containing temperature, precipitation and SPEI climate data for the period 1961-2015.}
\references{
  Rehschuh, R., Mette, T., Menzel, A., and Buras, A. (2017) \emph{Soil properties affect the drought susceptibility of Norway Spruce}.
  Dendrochronologia, 45, 81-89.
}
\keyword{datasets}