\name{SMPA}
\docType{data}
\alias{SMPA}
\title{Ring-width measurements of 40 Norway spruce specimen sampled south of Munich, Germany}
\description{
  This data set provides measurements of secondary growth of 40 Norway spruce specimen sampled
  south of Munich, Germany and was published in Rehschuh et al., 2017.
}
\usage{data(SMPA)}
\format{An rwl.data.frame containing ring-width measurements
 of 40 Norway spruce individuals from 1906-2015.}
\references{
  Rehschuh, R., Mette, T., Menzel, A., and Buras, A. (2017) \emph{Soil properties affect the drought susceptibility of Norway Spruce}.
  Dendrochronologia, 45, 81-89.
}
\keyword{datasets}